backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
pydantic[email]==1.10.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator>=2.0.0
alembic==1.12.0
psycopg2-binary>=2.9.1,<3.0.0
python-dotenv==1.0.0
pytest>=6.2.5,<7.0.0
httpx>=0.23.0,<0.24.0
pytest-cov>=2.12.0,<3.0.0
pytest-asyncio>=0.15.1,<0.16.0
python-json-logger==2.0.7